Fix layer mask selection (#24)
Unity internally removes "blank" layer fields, including (currently) three internal layers, which will change the value of any layer that is added after a blank layer. Using the to/from ConcatenatedLayersMask we can get the correct functionality of the layers mask selector.  (Small graphical bug: when "Everything" is selected, it will say "Mixed" because the blank layer fields will not be set.)